+++ examples/heptagram.rs
@@ -1,11 +1,11 @@
 // heptagram.rs
 extern crate footile;
 
-use footile::{FillRule, PathBuilder, Plotter, Transform};
+use footile::{FillRule,Gray8,PathBuilder,Plotter,Transform};
 
 fn main() -> Result<(), std::io::Error> {
     const PI: f32 = std::f32::consts::PI;
-    let mut p = Plotter::new(100, 100);
+    let mut p = Plotter::<Gray8>::new(100, 100);
     let h = (p.width() / 2u32) as f32;
     let q = h / 2f32;
     p.set_transform(Transform::new_scale(h, h).translate(q, q));