+++ examples/stroke.rs
@@ -1,7 +1,7 @@
 // stroke.rs
 extern crate footile;
 
-use footile::{PathBuilder, Plotter};
+use footile::{Gray8,PathBuilder,Plotter};
 
 fn main() -> Result<(), std::io::Error> {
     let path = PathBuilder::new().relative().pen_width(5f32)
@@ -9,6 +9,6 @@ fn main() -> Result<(), std::io::Error> {
                            .line_to(32f32, 0f32)
                            .line_to(-16f32, -32f32)
                            .close().build();
-    let mut p = Plotter::new(64, 64);
+    let mut p = Plotter::<Gray8>::new(64, 64);
     p.stroke(&path).write_png("./stroke.png")
 }